requirements.txt
rasa==2.0.0
python-jenkins
